Rip out Spectre.Console and replace it with Avalonia.FuncUI.
Spectre.Console is great, but it's still fundamentally a console app, which means parallelized interactivity is not supported. In this case we want the user to be able to (1) approve Dom5 versions for execution; (2) view the execution state of approved versions to see when they'll be done processing; (3) cleanup versions that are no longer interesting, either because they will never be approved or because they've already executed and been viewed in Dom5.